refactor: Add endpoints for managing admins
This commit adds new endpoints for managing admins in the super admin controller. The "getAdmins" endpoint retrieves a list of admins with pagination support, while the "toggleAdminStatus" endpoint allows toggling the status of an admin between active and inactive. These endpoints enhance the functionality of the application by providing the ability to view and manage admins.