refactor: Add putViewNews endpoint to super_admin.js router
The super_admin.js router has been updated to include a new endpoint, putViewNews, for increasing the views of a news article. This endpoint takes a query parameter, id, to specify the news ID. If the news exists and the user's ID is not already included in the viewers array, the user's ID is added to the viewers array and the news views are increased. This refactor enhances the functionality of the super admin section by providing the ability to track and increase news views.