refactor: Update .vscode/settings.json to customize colors
This commit updates the .vscode/settings.json file to customize the colors of the activity bar and title bar. The background color of the activity bar is set to #312E14, and the active background and foreground colors of the title bar are set to #45401B and #FBFAF4 respectively. This change improves the visual appearance of the application.