refactor: Update swagger.js and super_admin.js
This commit updates the swagger.js file to use the latest version of the swagger-jsdoc package. It also updates the super_admin.js file to remove the commented-out code related to the /superadmin/api/login endpoint. These changes ensure that the API documentation is up to date and the codebase is clean.