refactor: Add /send-message endpoint to send messages to all connected clients
This commit adds a new endpoint, /send-message, to the application. This endpoint allows sending a message to all connected clients using WebSocket. The endpoint expects a JSON payload with a "message" property. If the message is not provided, a 400 error is returned. Upon successful sending of the message, a 200 response is returned with a success message. This change enhances the functionality of the application by enabling real-time communication with all connected clients.