refactor: Update deleteSuperAdmin function to disallow deleting super admin accounts
This commit updates the deleteSuperAdmin function in super_admin.js to disallow deleting any super admin account except for the default super admin account with the email "lalkhan@superadmin.com". Previously, any admin account could be deleted, which posed a security risk. With this change, an error is thrown if an attempt is made to delete a super admin account other than the default one. This improves the security and integrity of the application.